+++ src/ui/DNA/pages/UpgradeDnaFlow/UpgradeDnaFlow.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Route, Routes } from 'react-router-dom';
-import { PageLayout } from 'src/ui/components/PageLayout';
+import { PageLayout } from 'src/ui/features/onboarding/shared/PageLayout';
 import { WideScreen } from '../../shared/WideScreen';
 import { Success } from './Success';
 import { SelectBackground } from './SelectBackground';
