+++ src/ui/features/onboarding/FAQ/FAQ.tsx
@@ -9,13 +9,13 @@ import { TextAnchor } from 'src/ui/ui-kit/TextAnchor';
 import { apostrophe } from 'src/ui/shared/typography';
 import { useWindowSizeStore } from 'src/ui/shared/useWindowSizeStore';
 import lockIconSrc from 'jsx:src/ui/assets/lock.png';
-import { SidePanel } from 'src/ui/components/SidePanel';
 import { Stack } from 'src/ui/ui-kit/Stack';
 import keyIconSrc from '../assets/key.png';
 import dialogIconSrc from '../assets/dialog.png';
 import metamaskIconSrc from '../assets/metamask.png';
 import metamaskInstructionSrc from '../assets/metamask_instruction.png';
 import walletIconSrc from '../assets/wallet2.png';
+import { SidePanel } from '../shared/SidePanel';
 import * as styles from './styles.module.css';
 
 export function SecretKeyFAQ() {
