+++ src/ui/features/onboarding/shared/SidePanel/SidePanel.tsx
