+++ src/ui/features/onboarding/shared/SidePanel/index.ts
