+++ src/ui/features/onboarding/shared/SidePanel/styles.module.css
